+++ src/main/java/me/towdium/jecalculation/network/ProxyServer.java
@@ -5,7 +5,6 @@ import me.towdium.jecalculation.JecaCapability;
 import me.towdium.jecalculation.data.label.ILabel;
 import me.towdium.jecalculation.data.structure.Recipes;
 import net.minecraftforge.common.capabilities.CapabilityManager;
-import net.minecraftforge.fml.relauncher.Side;
 
 /**
  * Author: towdium
@@ -20,10 +19,5 @@ public class ProxyServer implements IProxy {
     @Override
     public void initPost() {
         ILabel.initServer();
-    }
-
-    @Override
-    public void runOnSide(Runnable r, Side s) {
-        if (s == Side.SERVER) r.run();
     }
 }